+++ CHANGELOG.md
@@ -48,6 +48,8 @@ and this project adheres to
 - cosmwasm-std: Remove the previously deprecated `InitResult`, `HandleResult`,
   `MigrateResult` and `QueryResult` in order to make error type explicit and
   encourage migration to custom errors.
+- cosmwasm-std: Add a `data` field to `InitResponse` the same way as in
+  `MigrateResponse` and `HandleResponse`.
 - cosmwasm-vm: Avoid serialization of Modules in `InMemoryCache`, for
   performance. Also, remove `memory_limit` from `InstanceOptions`, and define it
   instead at `Cache` level (same memory limit for all cached instances).